here is one another thing my brother
i want that when someone clicks on each card or box then open a new page where user will be asked for status for that particular complaint which have two options "Pending" and "Completed" in select option manner. where user can select whether that particular task is completed or not if user selects "Pending" update it to databse with "Pending" text and if user selects that "Completed" then update it to databse that task is completed.

add and extra input field for comment above that select option menu and take inout from user and make it optional. whatever user enters it you have submit that text or input in databse.

make sure to add a submit button and whenevr user clicks on sub,it button then perform databse tasks