app is taking more time to load the page.
dont check if device is android or not just open it no matter whether user open is from laptop, android or ios or from web and theres one problmle solved it.

also remeber that when user add any comment in comment section then update it to databse and also remember that update all details to databse in lowercase form no matter if it is status or comment.

there's one thing more which is make sure that android or ios back button also work here they dont need to press abive back button always to go back to all complaints .  also chnage colour of cards with relevant coliur that are pending in action