time is in 24 hour format change it to 12 hour format
when i click on card then it's taking time fix it and make sure that it takes lesser time to open.

one more thing i want is that

first of all i wanna thank you brother  you helped me a lot.

now i want is that

when comment and status us fetch from databse and there are not changes in inputs of comment and status field then make submit button inaction and chnage the colur of submit button light blue colur that will represrnt that button isinacrive